requirements.txt
flask
flask-sqlalchemy
flask-admin
pillow
gunicorn
werkzeug
wtforms
